ck_hash: fix insert when hash_key starts at 0
When the initial hash_key is 0, the end is calculated by taking
the modules of -1 % table->capacity is rarely the end.
Adding table->capacity before taking the modulus fixes this.

I encountered a problem here when writing an SPSS file with the column
names as below. At position 'samendoen', the hash_key starts at 0,
and the derived end position is 3(!), and in this case, all 4 positions
are already occupied, so 'samendoen' is NOT inserted in the hash, and
the SPSS file will find equal short names for 'samendoen' and
'samendoen_partner' which yields strange results.

	resp
	id
	timezone
	posted
	modified
	created
	begin
	end
	mainact
	secact
	voldoening
	waar
	interactie_mrt18
	gesproken_niemand
	gesproken_collegas
	gesproken_externecontacten
	gesproken_partner
	gesproken_kind
	gesproken_ledengezin
	gesproken_uitfamilie
	gesproken_vriend
	gesproken_anderen
	gesproken_vrijwilligers
	kern_taak
	mainact_code_up1
	mainact_code_up2
	mainact_actclass
	secact_code_up1
	secact_code_up2
	secact_actclass
	begin_date
	begin_time
	begin_wkd
	end_date
	end_time
	end_wkd
	duration
	actnmb
	lastreg
	gesproken_werkcontacten
	gesproken_familie
	samenzijn
	samendoen
	alleen
	samenzijn_partner
	samenzijn_kind
	samenzijn_vriend
	samenzijn_familie
	samenzijn_werkcontacten
	samenzijn_anderen
	samendoen_partner
	samendoen_kind
	samendoen_vriend
	samendoen_familie
	samendoen_werkcontacten
	samendoen_collegas
	samendoen_vrijwilligers
	samendoen_externecontacten
	samendoen_anderen

Signed-off-by: Kurt Van Dijck <dev.kurt@vandijck-laurijssen.be>